pyproject.toml
[project]
name = "vercel-deploy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4==4.12.3",
    "fastapi==0.115.0",
    "groq==0.11.0",
    "httpx==0.27.2",
    "lxml==5.3.0",
    "mangum==0.17.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "supabase==2.9.1",
]
